+++ packages/tswebextension/src/lib/mv2/background/cosmetic-api.ts
@@ -21,6 +21,13 @@ type ScriptsAndScriptletsDataMv2 = {
     scriptText: string;
 };
 
+type LogJsRulesParamsMv2 = LogJsRulesParams & {
+    /**
+     * Cosmetic result.
+     */
+    cosmeticResult: CosmeticResult;
+};
+
 /**
  * Cosmetic api class.
  * Used to prepare and inject javascript and css into pages.
@@ -212,13 +219,16 @@ export class CosmeticApi extends CosmeticApiCommon {
      *
      * See {@link WebRequestApi.onBeforeRequest} for details.
      *
+     * TODO: Since injection and logging are happened in different places,
+     * we do not track was injection successful or not - we log all rules which
+     * expected to be applied. We should track injection result.
+     *
      * @param params Data for js rule logging.
      */
-    public static logScriptRules(params: LogJsRulesParams): void {
-        super.logScriptRules(
-            params,
-            CosmeticApi.shouldSanitizeScriptRule,
-        );
+    public static logScriptRules(params: LogJsRulesParamsMv2): void {
+        const scriptRules = CosmeticApi.filterScriptRulesForLog(params);
+
+        super.logScriptRules(params, scriptRules);
     }
 
     /**
@@ -269,6 +279,19 @@ export class CosmeticApi extends CosmeticApiCommon {
                 logger.debug('[tsweb.CosmeticApi.applyJs]: error occurred during injection', e);
             }
         }
+    }
+
+    /**
+     * Filters script rules for logging.
+     *
+     * @param params Data for JS rule logging.
+     *
+     * @returns Script rules which expected to be applied and logged.
+     */
+    private static filterScriptRulesForLog(params: LogJsRulesParamsMv2): CosmeticRule[] {
+        const scriptRules = params.cosmeticResult.getScriptRules();
+
+        return CosmeticApi.sanitizeScriptRules(scriptRules);
     }
 
     /**
@@ -336,3 +359,7 @@ export class CosmeticApi extends CosmeticApiCommon {
         }
     }
 }
+
+const cosmeticApi = new CosmeticApi();
+
+export { cosmeticApi };