+++ packages/tswebextension/CHANGELOG.md
@@ -24,6 +24,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Export `isUserScriptsApiSupported` function to check if the current browser
   supports User Scripts API.
 
+## [3.2.11] - 2025-09-11
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.2.3`.
+- Updated [@adguard/scriptlets] to `v2.2.10`.
+- Updated [@adguard/tsurlfilter] to `v3.4.6`.
+
+[3.2.11]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.2.11
+
+## [3.2.10] - 2025-09-09
+
+### Fixed
+
+- Mismatches in caching strategy for rulesets in MV3.
+
+[3.2.10]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.2.10
+
+## [3.2.9] - 2025-09-02
+
+### Fixed
+
+- Incorrect handling internal error message when passed empty array of user rules.
+
+[3.2.9]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.2.9
+
+## [3.2.8] - 2025-09-02
+
+### Removed
+
+- Static getter `isUserScriptsApiSupported`.
+
+### Added
+
+- Separate export `isUserScriptsApiEnabled` function from
+  `@adguard/tswebextension/mv3/utils` to check if the current browser supports
+  User Scripts API.
+
+[3.2.8]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.2.8
+
 ## [3.2.7] - 2025-07-30
 
 ### Changed